Add tests for ceil and expand
